+++ metadata.rb
@@ -12,8 +12,8 @@ version '1.0.0-pre'
   supports os
 end
 
-depends 'chef-sugar', '>=1.2.6'
 depends 'composer'
+depends 'ingenerator-helpers', '~> 1.0'
 # 1.7.0 introduces a breaking change regarding the quoting of attribute values - prevent updating
 # past 1.6.x until we have implemented some wrapping/warning and/or introduced a breaking
 # version of this cookbook to prevent unexpected production php config changes